pyproject.toml
[project]
dependencies = [
  "asyncpg>=0.31.0",
  "fastapi>=0.128.0",
  "jinja2>=3.1.6",
  "nanoid>=2.0.0",
  "pgvector>=0.4.2",
  "pydantic-ai>=1.44.0",
  "pyrefly>=0.48.2",
  "sqlmodel>=0.0.31",
]
description = "Add your description here"
name = "simple-bot-py"
readme = "README.md"
requires-python = ">=3.10"
version = "0.1.0"

[dependency-groups]
dev = [
  "ruff>=0.14.13",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
ignore = [
  "E501", # line too long (manejado por el formateador)
  "ANN101", # self no necesita type hint
  "ANN102", # cls no necesita type hint
]
select = [
  "E", # pycodestyle errors
  "F", # pyflakes
  "I", # isort
  "UP", # pyupgrade (sintaxis moderna)
  "B", # bugbear (mejores prácticas)
  "C4", # comprensiones
  "SIM", # simplificaciones
  "RUF", # reglas específicas de Ruff
  "N", # nomenclatura PEP8
  "ANN", # anotaciones de tipos obligatorias
  "S", # seguridad (bandit)
  "PTH", # usar pathlib en lugar de os.path
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"] # imports sin usar en __init__ están bien

[tool.ruff.format]
indent-style = "space"
quote-style = "double"
